consolidate CmdStubber into test package
