+++ command/testing.go
@@ -3,7 +3,6 @@ package command
 import (
 	"errors"
 	"fmt"
-	"os/exec"
 	"reflect"
 
 	"github.com/AlecAivazis/survey/v2"
@@ -11,42 +10,7 @@ import (
 
 	"github.com/cli/cli/api"
 	"github.com/cli/cli/context"
-	"github.com/cli/cli/internal/run"
-	"github.com/cli/cli/test"
 )
-
-// TODO this is split between here and test/helpers.go. I did that because otherwise our test
-// package would have to import utils (for utils.Runnable) which felt wrong. while utils_test
-// currently doesn't import test helpers, i don't see why that ought to be precluded.
-// I'm wondering if this is a case for having a global package just for storing Interfaces.
-type CmdStubber struct {
-	Stubs []*test.OutputStub
-	Count int
-	Calls []*exec.Cmd
-}
-
-func initCmdStubber() (*CmdStubber, func()) {
-	cs := CmdStubber{}
-	teardown := run.SetPrepareCmd(createStubbedPrepareCmd(&cs))
-	return &cs, teardown
-}
-
-func (cs *CmdStubber) Stub(desiredOutput string) {
-	// TODO maybe have some kind of command mapping but going simple for now
-	cs.Stubs = append(cs.Stubs, &test.OutputStub{[]byte(desiredOutput)})
-}
-
-func createStubbedPrepareCmd(cs *CmdStubber) func(*exec.Cmd) run.Runnable {
-	return func(cmd *exec.Cmd) run.Runnable {
-		cs.Calls = append(cs.Calls, cmd)
-		call := cs.Count
-		cs.Count += 1
-		if call >= len(cs.Stubs) {
-			panic(fmt.Sprintf("more execs than stubs. most recent call: %v", cmd))
-		}
-		return cs.Stubs[call]
-	}
-}
 
 type askStubber struct {
 	Asks  [][]*survey.Question
